Adds CurrentUser processor
This extracts the "find user by TGT" logic that exists in all of the processors
and puts it in a CurrentUser processor.

This processor is automatically run in a :before_filter for most controller
actions. If a user is found, then the User record is assigned to the
@current_user instance variable. From there it can be passed into the other
processors as well as be used in the various CASino views.

An additional benefit of this feature is that the host application can also
use this value in its own views:

```ruby
class WidgetsController < ApplicationController
  include CASino::CurrentUserHelper

  def index
    # ...
  end
end

<div class="widgets">
  <% if user_signed_in? %>
    <%= debug current_user %>
  <% end %>
</div>
```

* Adds CurrentUser processor to encapsulate all user look-up based on TGT
  values

* Provides #current_user and #user_signed_in? controller helper methods

* Refactored some more complicated internal logic out into private methods
  for a handful of processors

+++ app/processors/casino/processor_concern/ticket_granting_tickets.rb
@@ -3,8 +3,7 @@ require 'addressable/uri'
 module CASino
   module ProcessorConcern
     module TicketGrantingTickets
-
-      include CASino::ProcessorConcern::Browser
+      include Browser
 
       def find_valid_ticket_granting_ticket(tgt, user_agent, ignore_two_factor = false)
         ticket_granting_ticket = CASino::TicketGrantingTicket.where(ticket: tgt).first
@@ -50,10 +49,13 @@ module CASino
       end
 
       def remove_ticket_granting_ticket(ticket_granting_ticket, user_agent = nil)
-        tgt = find_valid_ticket_granting_ticket(ticket_granting_ticket, user_agent)
-        unless tgt.nil?
-          tgt.destroy
+        tgt = if ticket_granting_ticket.is_a? CASino::TicketGrantingTicket
+          ticket_granting_ticket
+        else
+          find_valid_ticket_granting_ticket(ticket_granting_ticket, user_agent)
         end
+
+        tgt.destroy if tgt
       end
 
       def cleanup_expired_ticket_granting_tickets(user)
